requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
SQLAlchemy
psycopg2-binary
pandas
llama-index
openai
tiktoken